// src/app/interfaces/AcolytePostion.ts
export enum AcolytePosition {
  ACOLYTE1 = "acolyte1",
  ACOLYTE2 = "acolyte2",
  MASTER_OF_CEREMONIES = "master_of_ceremonies",
  THURIFER = "thurifer",
  CRUCIFER = "crucifer",
  BOAT_BEARER = "boat_bearer",
}

export enum AcolytePositionTranslation {
  acolyte1 = "Acólito 1",
  acolyte2 = "Acólito 2",
  master_of_ceremonies = "Cerimoniário",
  thurifer = "Turiferário",
  crucifer = "Cruciferário",
  boat_bearer = "Porta-naveta",
}